fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,26 +12,26 @@ int main(void)
 
 void f(int n)
 {
-	long long n1 = 1;
+	unsigned long n1 = 1;
 
-	long long n2 = 2;
+	unsigned long n2 = 2;
 
 	int i;
 
-	printf("%llu, ", n1);
-	printf("%llu, ", n2);
+	printf("%lu, ", n1);
+	printf("%lu, ", n2);
 
 	for (i = 0; i < n - 2; i++)
 	{
-		long long n3 = n1 + n2;
+		unsigned long n3 = n1 + n2;
 
 		n1 = n2;
 		n2 = n3;
 
 		if (i == n - 3)
-			printf("%llu", n2);
+			printf("%lu", n2);
 		else
-			printf("%llu, ", n2);
+			printf("%lu, ", n2);
 	}
 	printf("\n");
 }
